redistribute utils.py
